Create LICENSE
Article (http://doi.org/10.1007/s10336-016-1404-9) and Zenodo package (http://doi.org/10.5281/zenodo.48423) mention this license for the source code, but the license is not added to the repository. Adding a LICENSE file is the most common method and interacts well with GitHub: license info will be shown in the header.